[PIX] Do not crash if the frame is not active.
Before this patch, every inactive main frame triggered a CHECK() in the
ContentFacilitatedPaymentsDriverFactory class.

This patch replaces the CHECK() with a graceful `if(!active)return`.

After this patch, if a frame is not active, then PIX code detection will
not be run and there will not be a crash.

Bug: 326531940
Change-Id: Ifd1791bc9f497544c80f4d36f51e65ccd4314d86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321438
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Sparik Hayrapetyan <ust@google.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264546}