Stop recording Signin.Signin.Started for reauth events
Clicking on "Sign in again" button while in the Sync paused state
eventually triggers `SigninViewController::ShowDiceSigninTab()` with
`signin_reason` set to kSigninPrimaryAccount. In its turn, this leads to
Signin.Signin.Started UMA being recorded at [1].

This is incorrect as Signin.Signin.Completed is never recorded at the
end of this flow since no new account is added to Chrome in the result.

To fix this, this CL changes the `signin_reason` to
`kReauthentication` if `ShowDiceAddAccountTab()` is called with an
email hint that points at an account that is known to the
`IdentityManager`. The user can still change this account on the Gaia
sign-in page which will make Signin.Signin.Started miss a
corresponding event. We believe that happens relatively rare.

As part of this change, reauthentication flows will start using the
regular sign-in URL instead of the Chrome Sync URL since the sync
prompt is never displayed in reauth flows [2].

[1] https://crsrc.org/c/chrome/browser/signin/dice_tab_helper.cc;drc=f4a00cc248dd2dc8ec8759fb51620d47b5114090;l=117
[2] https://crsrc.org/c/chrome/browser/signin/dice_tab_helper.cc;drc=28bad3b5120911c529daebdf42cc69167dc8f55c;l=107

Fixed: 325102895
Change-Id: I6ac72828db95196312ef7c7d0f97178e306be7fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307399
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264566}