webview: Enable guest to respect owner's muted state
This commit enables the guest WebContents to respect the owner
WebContents' muted state. This is done by checking the owner's muted
state prior to setting the muted state for the guest, and only unmuting when the owner is also unmuted. It uses the existing GuestViewBase's
owner WebContents observer to detect the changes in audio muted state.

Bug: b/40191772, b/325284151, b/40847159
Change-Id: I82a8c9e2b093fcd2a5df3fdf9f8bfea49cd35588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5296783
Reviewed-by: Chase Phillips <cmp@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Ovidio Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264578}