[FedCM] Reject button flow request without delay
The button flow is on a critical path so RPs should see the failures as
soon as possible to fallback to other sign-in mechanism. In addition,
the button flow does NOT have the same risk of leaking IdP status as the
widget flow does. e.g. when an RP doesn't get an immediate response, it
cannot differentiate the following two cases:
  1. if a user doesn't have an account with the IdP, we show the "Sign
  in to IdP pop-up window" and reject the promise when the user closes
  it.
  2. if a user has an account with the IdP, we show the account UI and
  reject the promie if the flow fails afterwards.

Therefore we don't need to add random delays like we do in the widget
flow.

This patch also fixes an rp_mode_ assignment issue.

Bug: 41482138
Change-Id: Ieccc1f526b18540d0cda3d6b338e1d9a1843a565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318058
Commit-Queue: Yi Gu <yigu@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264609}