webauthn: serialise EnclaveManager state at the right time.
If `WriteState` was called while a write was already in progress, the
code would set a flag and start a new write once the current one
completed. But that would mean that it would grab the state when the
write completed, rather than the state at the time at the write was
requested.

That should be fine, but it's a bit of non-determinism that feels a
little sharp, so change it to always capture the state at the time that
`WriteState` was called.

Change-Id: Ie31da1f6a26bbd6b4810b63e6c1b2a8276b21fef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5310034
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264618}