Log CheckDelta when response status is SuccessWithRealTime
Since GMSCore is still rolling out, a lot of real time checks fall
back to local blocklist checks. Add a histogram that logs CheckDelta
only if the response status is SuccessWithRealTime so that we get
a clear picture of the latency of real time check.

Bug: 40911907
Change-Id: I53ee1b2289744a6f40789c28c7a72e6168a571b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319779
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264634}