Move PDFiumEngine::LoadDocumentMetadata() impl into standalone function
Refactor existing code buried inside PDFiumEngine into a standalone
GetPDFiumDocumentMetadata() function that returns a struct
DocumentMetadata. This makes it possible for PDFiumEngineExports to call
it in a follow-up CL without having to repeat some of the same code.

This also reduces the amount of code in PDFiumEngine, and makes the code
that was in LoadDocumentMetadata() easier to unit test. To make the code
compile, change DocumentMetadata to be movable.

Bug: 40864319
Change-Id: Ib43151725465264cea43ae414ddd02f41bfc7f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319146
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264643}