[dPWA] Make the WebAppInstallDialogCoordinator be a WidgetObserver
Instead of using a raw_ptr to a bubble dialog delegate and then using
a WeakPtr to itself everywhere inside the view, the install dialog
coordinator for web apps is now a WidgetObserver. This ensures:
1. Simple lifetimes, since the view inside the coordinator is
tracked only as long as the WidgetObserver is observing the main
dialog, which is more robust than the previous implementation of
a criss-cross raw_ptr, weak_ptr interface.
2. Practices separation of concerns, and brings the implementation
closer to a MVC model, where the behavior, view and its observation
happens independent of each other, with no overlapping logic.

As a side note, this also fixes the measurement of the
WebApp.InstallConfirmation.CloseReason by moving it to a central
location to be used by both the old and new install dialogs.

Bug: b/326418546
Change-Id: I7672d82d1683aa45ce9e51a37d3fef2e1d0c0a22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320614
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264651}