[AriaNotify] Refactor AriaNotification
Previously, CL:5316723 updated the options for `ariaNotify`. This change
refactors `AriaNotification` to map these options to their respective
attributes in `AXNodeData`, which will be serialized in a later patch.

This change also drops the stored reference to the node that triggered
the notification to avoid using a garbage-collected vector.

Bug: 326277796
Change-Id: Iee1c95d55b119db96937aec6cab9a04f3710e9f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5316723
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1264652}