[picture-in-picture] Disable permission prompt buttons when occluded
Currently, permission prompts check whether a picture-in-picture window
occludes the button row before allowing keyboard input to click a
button. However, there are still ways in which this can be abused by a
nefarious website.

This CL strengthens that check to not allow any input to click a button
as long as any of the prompt is occluded by a picture-in-picture
window.

Bug: 40064754, 40069864
Change-Id: I212b98838d5242982db483cd77c63625ff5a1a0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5315611
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264670}