Remove special DetachLayoutTree for <select>
HTMLSelectElement::ChangeRendering has done a synchronous
DetachLayoutTree for a long time, and according to comments
and previous CLs (CL:2778391) some code that updated the
current selection index relied on this synchronous detach.
It's unclear what changed, and when, but we don't seem to
need this anymore.

TEST=HTMLSelectElementTest.ChangeRenderingCrash

Fixed: 40756551
Change-Id: Ib467a4d702fb6f9a8951f56e8d397f54816cfb6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5310041
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264671}