[BREFs] Fix error when mixing meta/header tokens
BREFs doesn't support header tokens, but this means we failed to test
what happens when meta and header tokens are mixed. As it turns out,
what happens it that the origin_trial_state_host_remote_ that's bound
when processing header tokens expires and future calls fail. We can
prevent this by checking if the remote isn't connected and resetting it
if so.

Bug: 326458800, 40243430
Change-Id: I2fd0f51466cdbef0923db14ef71f534047c54546
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318841
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264741}