wpt: Enable reftesting a prerendered page
Using reftest-wait in a prerendered page currently doesn't work since
notifyDone for this case is called from script installed by
TestRunnerBindings::Install only for the main window of a test. This
change marks a prerendering page as being in the main window since it
may be activated to become the main window. This ensures its initialized
appropriately and instead ensures the active main frame is used when
looking for _the_ main frame of the main window.

The script to watch for reftest finished, now installed in a prerender,
is amended to only watch for finish when the page becomes active.

A test relying on this is added in https://crrev.com/c/5314213.

Fixed: 40257783
Change-Id: I5eedadb46b10f33535579aa444b3eaf0ac152cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5317919
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264748}