Audio: Always choose hardware layout as output layout
When we playback a 4.0 channel layout audio, and when we have
a 5.1 channel layout Audio MIDI setup, their should be an upmix:

L -> L,
R -> R,
C -> C,
BC -> upmix to SL and SR

In this case, we expect the output channel layout always be 5.1,
but the previously logic limited us to use 4.0 layout, which
causes the silent back center channel.

To fix the bug, we should always use hardware layout as output
layout no matter what.

Bug: 40809490
Bug: 41495709
Bug: 41487675
Change-Id: If9f0ad0144c83a722622ea0e6f51286b9d8fc8f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320653
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Sida Zhu <zhusida@bytedance.com>
Cr-Commit-Position: refs/heads/main@{#1264756}