birch: Add initial BirchRanker with calendar and weather support
Introduce a BirchRanker, used by BirchModel to assign ranking scores to
items in the model. It's a separate class to keep the concerns separate
between the model (storage) and the ranker (computation). Only rank
calendar and weather for now to keep the CL size reasonable.

Make --force-birch-fetch dump the fetched items to the logs and add the
ranker scores to the logs. This allows manual debugging.

Bug: b:305094126
Test: added to ash_unittests
Change-Id: Ibaec31ec4b11963d3c2cdca330b2759ae84ca2f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319634
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264765}