Support FILE_CONTENTS for drop
Wayland currently supports FileContents type for drag, such as dragging
an image from a page. The file is offered with mime type
application/octet-string;name=<filename>.
https://source.chromium.org/chromium/chromium/src/+/main:ui/ozone/platform/wayland/host/wayland_exchange_data_provider.cc;l=212-223;drc=f4a00cc248dd2dc8ec8759fb51620d47b5114090

This adds support for receiving data in the same format and adding it to
OSExchangeData as FileContents, which allow images to be dragged from
an ash window (or some other client) and dropped into lacros.

Manual test:
* Open lacros at any "clipboard inspector" web page
* Open ash on any page with an image
* Drag image from ash to lacros and validate that image is seen

Bug: 1202034
Test: ozone_unittests, manual test
Change-Id: I4d5e5e214d2764ed9383ab2ec6cd9d8678cd26ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320629
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264773}