webauthn: pull the EnclaveManager state machine out.
When the primary user changes, the state machine should stop
immediately. Otherwise, if it continues, it could end up performing
actions on the wrong user account. While the account changing in the
middle of an action will be very rare, it is possible, esp if some
network service is being slow.

However, simply moving the state machine back to the idle state isn't
sufficient because Events from the previous execution might still be
inbound and need to be cleaned up. Each state could check an
`is_doomed_` flag and handle the clean up, but that seems fragile.

Instead, we want to use the `WeakPtr`s which are used with every Event.
That means that we need to destroy the state machine, but we don't want
to destroy the whole EnclaveManager. Thus the state machine needs to be
its own object.

Then, if the user account changes, we can simply delete the state
machine object and the `WeakPtr`s will take care of cleaning up any
in-flight Events.

Change-Id: I5e8720eee685b2a672099bc288a34fb5e7258dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5311064
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264774}