viz: win: Add feature to allow overlay promotion of non-scanout resources
This change introduces a Win-only feature,
`CopyNonOverlayResourcesToDCompSurfaces`. This allows
SkiaOutputDeviceDComp to display overlay candidates with resources that
are not scanout by copying them to a DComp surface before adding them
to the DComp tree. This is mainly meant to test delegated compositing.

This functionality will not be used in this CL, even with the feature
enabled. It requires OverlayProcessorWin changes coming with delegated
compositing.

Bug: 40150597
Change-Id: I2bc77e3182a12e0ea19b446cfe264ad70e8b6d8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5281136
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1264805}