snap-group: Fix move group item to another desk crash with bubble
This cl fixed the crash when dragging a group item to another desk with
the existence of bubble widget. The `overview_item_` associated with the
`OverviewItemView` may be reset during window restoring.

To prevent the cyclic calls to access the `overview_item_`, we should
not set bounds on the overview item during the restore process.

Fixed: b/311255082
Demo: http://b/311255082#comment5
Test: Added new unit test(verified w/wo fix) + manual
Change-Id: If45917826477e8ac83a5ce56cfdf9666a1fd3812
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320070
Commit-Queue: Michele Fan <michelefan@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264826}