[PWA] Deflake PWA install views browser tests on Mac
This CL attempts to deflake the PWA install views browser tests on Mac
by explicitly observing and waiting for the destruction of the
PWA installation dialog instead of a blanket ExecuteForTesting() call.
This ensures that tests wait for the dialog to fully close which then
synchronously updates the install view before running assertions.

This CL also renames some functions in the test file for readability.

Bug: 326554287
Change-Id: I6a976adc125c6697694e8022d5b56a1037b354b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5322461
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264834}