Add metric for wallpaper search session setting theme
This adds an enum metric that stores the type of theme set by wallpaper
search, if it is still the theme when the customize chrome side panel
is closed. Inspiration isn't saved in history so it can't take
advantage of the result from SaveCurrentBackgroundToHistory() to know if
it is still the current on session end. Therefore, if an inspiration
image was set in the session, it will check if it is the set one and
save it to the histogram.

Bug: 324648372
Change-Id: Ifa868ebee0439f3aaa7d1c0db049ccb3db16c35e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318581
Commit-Queue: Riley Tatum <rtatum@google.com>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264835}