Prefetch weight files before loading on-device ML model
This attempts to prefetch the weight files from disk when loading the
model, which can improve performance quite a bit when the weights are
not in the file cache. The weight files for the ML model can be >1 GB,
so reading from disk is a significant piece of startup time.

When testing on Linux, using POSIX_FADV_SEQUENTIAL gives much better
perf than WILLNEED. A follow up patch will add the ability to do this
in PreReadFile.

On my Windows desktop this improves loading time when the file is not
cached by about 500ms. Linux and Mac also have similar magnitude of
improvement.

Bug: 326240401
Change-Id: I916fd8280a98023184cc2d63fb4b71ccc2539ae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5313904
Auto-Submit: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264842}