[SHv2] Fix backwards incompatibility with stored results
As part of crrev.com/c/5000387, the result that is stored in prefs was
simplified and minimized. For the result of unused site permissions,
this also resulted in the format of the stored Dict to be changed. As a
result, when updating, users would still have the older version of the
stored Dict while the code was expecting the new version, resulting in a
crash. This CL addresses this issue by checking whether the the data
format is still of an older version, or whether it is of a new one.

Bug: 326600383, 40267370
Change-Id: I2840d227fd1e016ab667bc0b0af757615767c51c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321247
Reviewed-by: Side YILMAZ <sideyilmaz@chromium.org>
Commit-Queue: Tom Van Goethem <tov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1264993}