arc: Don't send unused power message to ARCVM
ARCVM ignores PowerSupplyInfoChanged since it gets the necessary
information from the goldfish battery. If ARCVM is going to ignore the
message anyway, then it's better to not do the work of sending it to the
guest in the first place.

Bug: b:322919652
Change-Id: Ia8404b96f0467d40d2150957b9b7a45b3894d9e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5260029
Auto-Submit: David Stevens <stevensd@chromium.org>
Reviewed-by: Shengsong Tan <sstan@chromium.org>
Reviewed-by: Alexandre Marciano Gimenez <raging@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265025}