Remove tapted@chromium.org from chrome/test/base/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/tapted@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I15778051d0a6fa304bf4e4f18d0bc7ab079a6135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5291531
Reviewed-by: Trent Apted (tapted@chromium) <tapted@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265096}