[iOS] Update Grid Cell style in light mode
Dark mode is forced at the window level when the user is in the TabGrid.
To ensure we are respecting the user choice for the regular grid cell,
when the cell is moving to the window, take the original style of the
window.

Fixed: 326597273
Change-Id: I0344c2dfba6c484b041b10ee23045512af849af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5317668
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265102}