[bookmarks] Avoid passing BookmarkModel in observer APIs
Including a pointer to BookmarkModel in BookmarkModelObserver APIs is
somewhat redundant, as observers usually know what they are observing.
They often do so by either maintaining a pointer to BookmarkModel
directly or because they use a ScopedObservation (which also exposes
BookmarkModel via ScopedObservation::GetSource()).

In addition to being redundant, it is error-prone for ongoing efforts
(see linked bug) that reuse BookmarkModelObserver when dealing with
abstractions that expose a transformed or limited view of an underlying
BookmarkModel instance. This is the case for
sync_bookmarks::BookmarkModelView, which benefits directly in this patch
as it is obvious in the changes to class
sync_bookmarks::BookmarkModelObserverImpl, and analogous concepts are
planned to be introduced in future patches.

The patch is mostly mechanical as most observers either didn't use the
parameter or should trivially switch to something else. The few
exceptions with less trivial changes include:
1. BookmarkUndoService, which temporarily (and until the migration
   work that motivates this patch is completed) observes multiple
   BookmarkModel instances.
2. The iOS version of HistoryClientImpl, which now issues more lookups
   to determine whether a URL has, in fact, been un-bookmarked (instead
   of smartly issuing the lookup in the "other" BookmarkModel instance).
   These two lookups will also go away once the very same migration
   is completed.

Change-Id: I48c90b467562bd6c95f603e7f37a0ffa698f4048
Bug: 326185948
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318097
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265117}